Pull postFold() from functionResolution.cpp
